Extract `LoadingSkeleton` that can be reused between list and card view
